Moved manager.rs file contents to parent modules
